removed console logs, fixed formik default values, refactored duplicate AutoCompleteInput components namspace collision
