+++ client/src/components/forms/auto-complete-input.tsx
@@ -1,4 +1,4 @@
-import { FC } from 'react';
+import { FC, useRef, useEffect } from 'react';
 import { Autocomplete, FormControl, TextField, TextFieldProps } from '@mui/material';
 import { useFormikContext } from 'formik';
 
@@ -12,8 +12,53 @@ export interface AutoCompleteInputProps {
   options: AutoCompleteOption[];
 }
 
-export const AutoCompleteInput: FC<AutoCompleteInputProps & TextFieldProps> = (props) => {
+interface AutoCompleteInputPropsExt extends AutoCompleteInputProps {
+  tabSelection: number;
+}
+
+export const CompanyAutoCompleteInput: FC<AutoCompleteInputProps & TextFieldProps> = (props) => {
   const { handleChange, handleBlur, values, touched, errors, isSubmitting } = useFormikContext<any>();
+  const selectedOption = props.options.find(opt => opt.value === values[props.name]);
+  return (
+    <FormControl variant={props.variant} fullWidth={props.fullWidth}>
+      <Autocomplete
+        options={props.options}
+        renderInput={(params: any) => (
+          <TextField {...props} {...params} error={!!errors[props.name]} helperText={(touched[props.name] && errors[props.name]) as string} InputLabelProps={{ shrink: true }} />
+        )}
+        onChange={(event, option) => 
+          handleChange({
+            target: {
+              name: props.name,
+              value: option ? option.value : ''
+            }
+          })
+        } 
+        value={selectedOption}
+        getOptionLabel={(option) => option ? option.label : ''}
+        onBlur={handleBlur}
+        disabled={props.disabled || isSubmitting}
+      />
+    </FormControl>
+  );
+};
+
+export const TableAutoCompleteInput: FC<AutoCompleteInputPropsExt & TextFieldProps> = (props) => {
+  const { handleChange, handleBlur, values, touched, errors, isSubmitting, setFieldValue } = useFormikContext<any>();
+  const { tabSelection } = props;
+  const prevTabSelection = useRef(tabSelection);
+
+  useEffect(() => {
+    if (prevTabSelection.current !== tabSelection) {
+      const newValue = tabSelection === 1 ? 'small' : 'it';
+      setFieldValue(props.name, newValue);
+      prevTabSelection.current = tabSelection;
+    }
+  }, [tabSelection]);
+
+  const selectedValue = values[props.name];
+  const autoCompleteValue = props.options.find((option) => option.value === selectedValue);
+
   return (
     <FormControl variant={props.variant} fullWidth={props.fullWidth}>
       <Autocomplete
@@ -25,7 +70,7 @@ export const AutoCompleteInput: FC<AutoCompleteInputProps & TextFieldProps> = (p
           handleChange({ target: { name: props.name, value: value?.value } });
         }}
         onBlur={handleBlur}
-        value={values[props.name]}
+        value={autoCompleteValue}
         getOptionLabel={(option) => option.label}
         disabled={props.disabled || isSubmitting}
       />
