+++ client/src/components/session-decrypt/decrypt-input.tsx
@@ -82,9 +82,7 @@ export const DecryptInputForm = () => {
   // Can replace this function with any custom functionality to apply
   // over unencrypted shares.
   async function decrypt(pKey: string) {
-    console.log('privateKey received');
     if (token !== undefined && sessionId !== undefined) {
-      console.log('decryption started');
       const privateCryptoKey = await importPemPrivateKey(pKey);
       const { data, total_cells, metadata } = await getSubmissions(sessionId, token);
 
@@ -95,7 +93,6 @@ export const DecryptInputForm = () => {
 
       const scale = (num: number) => num / 100;
       const decodedTable = await secretSharesToTable(data, privateCryptoKey, bigPrime, reduce, recordProgress, scale);
-      console.log('decryption completed');
       dispatch(setDecodedTable(decodedTable));
       dispatch(setMetadata(metadata));
     } else {
