+++ client/src/pages/home.tsx
@@ -43,24 +43,16 @@ export const HomePage: FC = () => {
   useEffect(() => {
     const loadData = async () => {
       if (file) {
-        console.log('Starting MPC Encryption');
         const csvData = await readCsv(file);
-        console.log('Validating Data');
         setIsDataValid(validateData(csvData));
-        console.log('Loading File Data');
         setData(csvData);
         const scale = (num: number) => num * 100;
         const prime = await getPrime(sessionId);
-        console.log('Loading Prime Number');
         const publicKeyString = await getPublicKey(sessionId);
-        console.log('Loading Public Key');
         const publicCryptoKey = await importPemPublicKey(publicKeyString);
-        console.log('Imported Public Key');
         const secretTable = await tableToSecretShares(csvData, numShares, threshold, numEncryptWithKey, publicCryptoKey, new BigNumber(prime), true, scale);
-        console.log('Encrypted Data');
         setTable(secretTable);
         setDataIsEncrypted(Object.keys(data).length !== 0);
-        console.log('MPC Encryption Complete');
       }
     };
     loadData();
